+++ packages/astro/src/core/add/index.ts
@@ -705,8 +705,12 @@ export async function validateIntegrations(integrations: string[]): Promise<Inte
 				];
 
 				if (pkgJson['peerDependencies']) {
+					const meta = pkgJson['peerDependenciesMeta'] || {};
 					for (const peer in pkgJson['peerDependencies']) {
-						dependencies.push([peer, pkgJson['peerDependencies'][peer]]);
+						const optional = meta[peer]?.optional || false;
+						if (!optional) {
+							dependencies.push([peer, pkgJson['peerDependencies'][peer]]);
+						}
 					}
 				}