+++ packages/astro/src/vite-plugin-markdown/index.ts
@@ -5,9 +5,10 @@ import { fileURLToPath } from 'node:url';
 import type { Plugin } from 'vite';
 import { normalizePath } from 'vite';
 import type { AstroSettings } from '../@types/astro';
-import { collectErrorMetadata } from '../core/errors.js';
+import { AstroErrorCodes, MarkdownError } from '../core/errors/index.js';
 import type { LogOptions } from '../core/logger/core.js';
 import { warn } from '../core/logger/core.js';
+import { isMarkdownFile } from '../core/util.js';
 import type { PluginMetadata } from '../vite-plugin-astro/types.js';
 import { getFileInfo, safelyGetAstroData } from '../vite-plugin-utils/index.js';
 
@@ -19,9 +20,28 @@ interface AstroPluginOptions {
 function safeMatter(source: string, id: string) {
 	try {
 		return matter(source);
-	} catch (e) {
-		(e as any).id = id;
-		throw collectErrorMetadata(e);
+	} catch (err: any) {
+		const markdownError = new MarkdownError({
+			errorCode: AstroErrorCodes.GenericMarkdownError,
+			message: err.message,
+			stack: err.stack,
+			location: {
+				file: id,
+			},
+		});
+
+		if (err.name === 'YAMLException') {
+			markdownError.setErrorCode(AstroErrorCodes.MarkdownFrontmatterParseError);
+			markdownError.setLocation({
+				file: id,
+				line: err.mark.line,
+				column: err.mark.column,
+			});
+
+			markdownError.setMessage(err.reason);
+		}
+
+		throw markdownError;
 	}
 }
 
@@ -39,7 +59,7 @@ export default function markdown({ settings, logging }: AstroPluginOptions): Plu
 		// passing to the transform hook. This lets us get the truly raw value
 		// to escape "import.meta.env" ourselves.
 		async load(id) {
-			if (id.endsWith('.md')) {
+			if (isMarkdownFile(id, { criteria: 'endsWith' })) {
 				const { fileId, fileUrl } = getFileInfo(id, settings.config);
 				const rawFile = await fs.promises.readFile(fileId, 'utf-8');
 				const raw = safeMatter(rawFile, id);
@@ -63,7 +83,7 @@ export default function markdown({ settings, logging }: AstroPluginOptions): Plu
 					warn(
 						logging,
 						'markdown',
-						`[${id}] Astro now supports MDX! Support for components in ".md" files using the "setup" frontmatter is no longer enabled by default. Migrate this file to MDX or add the "legacy.astroFlavoredMarkdown" config flag to re-enable support.`
+						`[${id}] Astro now supports MDX! Support for components in ".md" (or alternative extensions like ".markdown") files using the "setup" frontmatter is no longer enabled by default. Migrate this file to MDX or add the "legacy.astroFlavoredMarkdown" config flag to re-enable support.`
 					);
 				}